+++ highdsa-service-mybatis/src/main/java/pers/husen/highdsa/service/mybatis/dao/customer/CustUserInfoMapper.java
@@ -4,11 +4,11 @@ import java.util.List;
 import pers.husen.highdsa.common.entity.po.customer.CustUserInfo;
 
 public interface CustUserInfoMapper {
-    int deleteByPrimaryKey(Integer userId);
+    int deleteByPrimaryKey(Long userId);
 
     int insert(CustUserInfo record);
 
-    CustUserInfo selectByPrimaryKey(Integer userId);
+    CustUserInfo selectByPrimaryKey(Long userId);
 
     List<CustUserInfo> selectAll();
